perf: u32 Dial bucket entries + u16 arc-cost storage
Dial bucket entries shrink from (node, queued_dist) 16-byte tuples to bare
u32 node ids. The queued distance was redundant: relaxations push
nd = cur_dist + rc with rc < k, and a bucket fully drains before the scan
advances past it, so every entry pops at exactly the distance it was queued
with - staleness reduces to `dist[u] != cur_dist` (plus the existing popped
check). Applies to the serial, full-completion (FIFO), and parallel Dial
variants and the single-source SSP. 4x less queue traffic in the hot loop.

Network::cost_fwd shrinks from i32 to u16, halving the largest per-frame
solver allocation (~4 arcs/pixel; SNAPHU likewise stores costs as short).
Every builder fits: the parity spline cost caps at 100*ln(1e30) ~= 6,908,
the analytical Carballo LUT at 300, and the CRLB / sparse inverse-variance
builders now saturate at the new cost::MAX_ARC_COST = 65,535 (an arc that
expensive - LLR ~655 - is already "never cut here"). Construction asserts
the range so an out-of-range cost builder fails fast instead of wrapping.

Validated on the 13-frame NISAR GUNW bench (baseline vs modified builds):
unwrapped phase AND conncomp byte-identical on all 13 frames; runtime
-11.6% mean (D_075 32.2s -> 27.3s; D_077 cumulative with the previous
slim-down 34.6s -> 26.6s); peak RSS -7.9% mean. 97/97 core tests green.

+++ crates/whirlwind-core/src/conncomp.rs
@@ -78,7 +78,7 @@ fn edge_is_cut(net: &Network, fwd1: usize, fwd2: usize, thresh: i32) -> bool {
     if forbidden1 || forbidden2 {
         return true;
     }
-    net.cost_fwd[fwd1].min(net.cost_fwd[fwd2]) <= thresh
+    (net.cost_fwd[fwd1].min(net.cost_fwd[fwd2]) as i32) <= thresh
 }
 
 /// Grow components on the pixel grid using a solved MCF network. Returns a
